requirements.txt
google-genai
python-dotenv
pandas
tqdm
scikit-learn